Changed wordings and adding banner to match new model
