+++ src/features/profile/pages/ChangeEmail/ChangeEmailContent.tsx
@@ -1,21 +1,25 @@
 import { yupResolver } from '@hookform/resolvers/yup'
-import React, { useEffect, useCallback } from 'react'
+import React from 'react'
 import { useForm } from 'react-hook-form'
 import { useSafeAreaInsets } from 'react-native-safe-area-context'
 
 import { UserProfileResponse } from 'api/gen'
 import { AlreadyChangedEmailDisclaimer } from 'features/profile/components/Disclaimers/AlreadyChangedEmailDisclaimer'
 import { ChangeEmailDisclaimer } from 'features/profile/components/Disclaimers/ChangeEmailDisclaimer'
-import { useChangeEmailMutation } from 'features/profile/helpers/useChangeEmailMutation'
 import { changeEmailSchema } from 'features/profile/pages/ChangeEmail/schema/changeEmailSchema'
 import { EmailInputController } from 'shared/forms/controllers/EmailInputController'
-import { PasswordInputController } from 'shared/forms/controllers/PasswordInputController'
+import { InfoBanner } from 'ui/components/banners/InfoBanner'
 import { ButtonPrimary } from 'ui/components/buttons/ButtonPrimary'
 import { Form } from 'ui/components/Form'
 import { SUGGESTION_DELAY_IN_MS } from 'ui/components/inputs/EmailInputWithSpellingHelp/useEmailSpellingHelp'
+import { Info } from 'ui/svg/icons/Info'
 import { Spacer } from 'ui/theme'
 
-import { FormValues, CenteredContainer, ButtonContainer } from './ChangeEmail'
+import { CenteredContainer, ButtonContainer } from './ChangeEmail'
+
+export type FormValues = {
+  currentEmail: string
+}
 
 export function ChangeEmailContent({
   hasCurrentEmailChange,
@@ -30,43 +34,25 @@ export function ChangeEmailContent({
 }) {
   const {
     control,
-    handleSubmit,
-    formState: { isValid },
-    setError,
-    watch,
-    clearErrors,
+    // handleSubmit,
+    // formState: { isValid },
+    // setError,
+    // watch,
+    // clearErrors,
   } = useForm<FormValues>({
     defaultValues: {
-      newEmail: '',
-      password: '',
+      currentEmail: '',
     },
     resolver: yupResolver(changeEmailSchema(user?.email)),
     mode: 'all',
     delayError: SUGGESTION_DELAY_IN_MS,
   })
-  const { changeEmail, isLoading } = useChangeEmailMutation({
-    setPasswordErrorMessage: (message: string) =>
-      setError('password', { message, type: 'validate' }),
-  })
 
-  const removePasswordError = useCallback(() => {
-    clearErrors('password')
-  }, [clearErrors])
-
-  const password = watch('password')
-
-  useEffect(() => {
-    removePasswordError()
-  }, [password, removePasswordError])
+  // TODO: useChangeEmailMutation & submitEmailChange
 
   const { bottom } = useSafeAreaInsets()
 
-  const submitEmailChange = ({ newEmail, password }: FormValues) => {
-    changeEmail({ email: newEmail, password })
-  }
-
-  const isInputDisabled = hasCurrentEmailChange
-  const isSubmitButtonDisabled = !isValid || isLoading
+  // const isSubmitButtonDisabled = !isValid || isLoading
   return (
     <React.Fragment>
       <Spacer.Column numberOfSpaces={6} />
@@ -82,20 +68,19 @@ export function ChangeEmailContent({
         <Form.MaxWidth flex={1}>
           <EmailInputController
             control={control}
-            name="newEmail"
-            label="Nouvelle adresse e-mail"
-            placeholder="email@exemple.com"
-            disabled={isInputDisabled}
+            name="currentEmail"
+            label="Adresse e-mail actuelle"
+            placeholder={user?.email}
+            disabled
             autoFocus
-            isRequiredField
           />
           <Spacer.Column numberOfSpaces={4} />
-          <PasswordInputController
-            control={control}
-            name="password"
-            disabled={isInputDisabled}
-            isRequiredField
+          <InfoBanner
+            icon={Info}
+            message="Tu vas recevoir un lien de confirmation sur ton adresse e-mail actuelle. Ce lien est valable 24h."
           />
+          <Spacer.Column numberOfSpaces={4} />
+
           {isMobileViewport && isTouch ? (
             <Spacer.Flex flex={1} />
           ) : (
@@ -105,10 +90,10 @@ export function ChangeEmailContent({
           <Spacer.Column numberOfSpaces={8} />
           <ButtonContainer paddingBottom={bottom}>
             <ButtonPrimary
-              wording="Valider la demande"
+              wording="Envoyer la demande"
               accessibilityLabel="Valider la demande de modification de mon e-mail"
-              onPress={handleSubmit(submitEmailChange)}
-              disabled={isSubmitButtonDisabled}
+              // onPress={handleSubmit(submitEmailChange)}
+              // disabled={isSubmitButtonDisabled}
             />
           </ButtonContainer>
         </Form.MaxWidth>
